Message payloads are binary, not text. v1.14.
Previously, I thought that every string in MQTT was UTF-8 encoded. That was
wrong: every string, except the message payload in a PUBLISH packet, is UTF-8,
but application messages themselves are raw binary and may include null bytes.
This was confirmed in the #mqtt IRC channel on Freenode.

This puts the burden of encoding and decoding on the user of this module, and
it means that generic applications like bin/mqtt-simple may have to deal with
mixed input of unknown encodings. I'm "solving" this by passing the message to
stdout as-is.

Additionally, I've replaced tabs with spaces in Changes.