Update LICENSE
This makes it explicit that a distributor is in fact allowed to reduce
the set of licenses it is distributed under.

When taken literally, this change means you may from now on also choose
OSI-approved licenses which you do not consider to be your favourites,
for instance when you expect that a certain subset is more likely to be
accepted by your policy enforcers.